restructure code
* split `mdb_attach.py` into 3 separate components:
  - `mdb_attach.py` for attaching to the `gdb` procs
  - `mdb_client.py` for managing the connections to the remote debugger
  - `mdb_shell.py` a class for the internal mdb command prompt
* refactor has removed some global vars
* I now use `isort` to order imports
